Add documentation for page-level params
There previously wasn't documentation for things such as disabling
comments on a page or removing the table of contents despite these
features existing. This commit documents all the page-level params I
could find as well as their defaults.

I've also updated the example config so that it shows all params.