Switch to a more flexible and consistent custom header
I previously introduced a change to add a custom stylesheet at the
bottom of the page, but this was actually far more limiting that it
needed to be. It only allowed the user to add their styles to a single
stylesheed which is not great if they wish to add a third party
stylesheet as well as their own styles.

This change also makes this addition consistent with the other custom
partial introduced in this pull request.